refactor sasl auth
Refactor broker code and move out SASL auth logic to its own
auth strategy and abstract authentication module out to make it more
extendable in the future.
