+++ dev/openresty/nginx.conf.d/main.conf
@@ -14,7 +14,7 @@ server {
                 local broker_list = {
                     { host = "broker", port = 9093 },
                 }
-                local sasl_config = { mechanism="PLAIN", user="admin", password="admin-secret" }
+                local auth_config = { strategy = "sasl", mechanism="PLAIN", user="admin", password="admin-secret" }
 
                 local key = "key"
                 local message = "hallo world"
@@ -46,7 +46,7 @@ server {
                     ssl = true,
                     client_cert = CERT,
                     client_priv_key = CERT_KEY,
-                    auth_config = sasl_config
+                    auth_config = auth_config
                 }
 
                 -- usually we do not use this library directly
