+++ lib/resty/kafka/broker.lua
@@ -3,7 +3,7 @@
 
 local response = require "resty.kafka.response"
 local request = require "resty.kafka.request"
-local sasl = require "resty.kafka.sasl"
+local auth_utils = require "resty.kafka.auth.init"
 local pid = ngx.worker.pid
 
 local to_int32 = response.to_int32
@@ -15,30 +15,13 @@ local _M = {}
 local mt = { __index = _M }
 
 
-function _M.new(self, host, port, socket_config, sasl_config)
+function _M.new(self, host, port, socket_config, auth_config)
     return setmetatable({
         host = host,
         port = port,
         config = socket_config,
-        auth = sasl_config or nil,
+        auth = auth_config or nil,
     }, mt)
-end
-
-local function sasl_auth(sock, broker)
-    local ok, err = _sasl_handshake(sock, broker)
-    if not ok then
-        if err then
-            ngx.say("sasl handshake failed -> See list of supported Mechanisms: " .. err)
-            return nil, err
-        end
-        return nil, "Unkown Error"
-    end
-
-    local ok, err = _sasl_auth(sock, broker)
-    if not ok then
-        return nil, err
-    end
-    return 0, nil
 end
 
 function _M.send_receive(self, request)
@@ -46,8 +29,6 @@ function _M.send_receive(self, request)
     if not sock then
         return nil, err, true
     end
-    -- TODO: figure out how this works
-    -- local times, err = sock:getreusedtimes()
 
     sock:settimeout(self.config.socket_timeout)
 
@@ -72,9 +53,16 @@ function _M.send_receive(self, request)
         end
     end
 
-    -- authenticate only if it is a new socket connection
+    -- authenticate if auth config is passed and socked connection is new
     if self.auth and sock:getreusedtimes() == 0 then -- SASL AUTH
-        local ok, err = sasl_auth(sock, self)
+        local ok, auth, err
+
+        auth, err = auth_utils.new(self.auth)
+        if not auth then
+            return nil, err
+        end
+
+        ok, err = auth:authenticate(sock)
         if not ok then
             local msg = "failed to do " .. self.auth.mechanism .." auth with " ..
                         self.host .. ":" .. tostring(self.port) .. ": " .. err, true
@@ -84,13 +72,13 @@ function _M.send_receive(self, request)
         ngx.say("Authentication successful")
     end
 
-    local data, err, f  = _sock_send_recieve(sock, request)
+    local data, err, f  = _sock_send_receive(sock, request)
     sock:setkeepalive(self.config.keepalive_timeout, self.config.keepalive_size)
     return data, err, f
 end
 
 
-function _sock_send_recieve(sock, request)
+function _sock_send_receive(sock, request)
     local bytes, err = sock:send(request:package())
     if not bytes then
         return nil, err, true
@@ -98,7 +86,7 @@ function _sock_send_recieve(sock, request)
 
     -- Reading a 4 byte `message_size`
     local len, err = sock:receive(4)
-    
+
     if not len then
         if err == "timeout" then
             sock:close()
@@ -117,85 +105,5 @@ function _sock_send_recieve(sock, request)
 
     return response:new(data, request.api_version), nil, true
 end
-
-
-function _sasl_handshake_decode(resp)
-    -- TODO: contains mechanisms supported by the local server
-    -- read this like I did with the supported api versions thing
-    local err_code =  resp:int16()
-    local mechanisms =  resp:string()
-    ngx.say("Decoding sasl handshake response -> ")
-    ngx.say("err_code -> " .. err_code)
-    ngx.say("mechanisms -> " .. mechanisms)
-    if err_code ~= 0 then
-        return err_code, mechanisms
-    end
-    return 0, nil
-end
-
-
-function _sasl_auth_decode(resp)
-    local err_code = resp:int16()
-    local error_msg  = resp:nullable_string()
-    ngx.say("sasl auth decode, error_code " .. err_code)
-    ngx.say("sasl auth decode, error_message " .. error_msg)
-    local auth_bytes  = resp:bytes()
-    ngx.say("sasl auth decode, auth_bytes " .. auth_bytes)
-    if err_code ~= 0 then
-        return nil, error_msg
-    end
-    return 0, nil
-end
-
-
-function _sasl_auth(sock, brk)
-    local cli_id = "worker" .. pid()
-    local req = request:new(request.SaslAuthenticateRequest, 0, cli_id, request.API_VERSION_V1)
-    local mechanism = brk.auth.mechanism
-    ngx.say("Authenticating with mechanism -> " .. mechanism)
-    local user = brk.auth.user
-    ngx.say("Authenticating with user -> " .. user)
-    local password = brk.auth.password
-    ngx.say("Authenticating with pwd-> " .. password)
-    local msg = sasl.encode(mechanism, user, password)
-    req:bytes(msg)
-    local resp, err = _sock_send_recieve(sock, req)
-    if not resp  then
-        ngx.say("Authentication failed with " .. err)
-        return nil, err
-    end
-    ngx.say("Authentication succeeded, decoding sasl_auth response")
-    local rc, err = _sasl_auth_decode(resp)
-    if not rc then
-        if err then
-            return nil, err
-        end
-        return nil, "Unkown Error during _sasl_auth"
-    end
-    return rc, err
-end
-
-
-function _sasl_handshake(sock, brk)
-    local cli_id = "worker" .. pid()
-    local api_version = request.API_VERSION_V1
-
-    local req = request:new(request.SaslHandshakeRequest, 0, cli_id, api_version)
-    local mechanism = brk.auth.mechanism
-    req:string(mechanism)
-    ngx.say("Requesting handshake for mechanism -> " .. mechanism)
-    local resp, err = _sock_send_recieve(sock, req)
-    if not resp  then
-        ngx.say("No response from sock_send_receive -> " .. err)
-        return nil, err
-    end
-    local rc, mechanism = _sasl_handshake_decode(resp)
-    -- the presence of mechanisms indicate that the mechanism used isn't enabled on the Kafka server.
-    if mechanism then
-        return nil, mechanism
-    end
-    return rc, nil
-end
-
 
 return _M
